+++ src/clj/lasso/config.clj
@@ -41,8 +41,13 @@
             :api-base-url (get-env "LASTFM_API_BASE_URL" "https://ws.audioscrobbler.com")
             :auth-url (get-env "LASTFM_AUTH_URL" "https://www.last.fm")
             :callback-url (get-env "OAUTH_CALLBACK_URL" "http://localhost:8080/api/auth/callback")}
-   :session {:secret (get-env "SESSION_SECRET" "development-secret-change-in-production")}
+   :session {:secret (get-env "SESSION_SECRET" "development-secret-change-in-production")
+             :ttl-ms (Long/parseLong (get-env "SESSION_TTL_MS" "86400000"))              ; 24h
+             :cleanup-interval-ms (Long/parseLong (get-env "SESSION_CLEANUP_INTERVAL_MS" "3600000"))} ; 1h
    :polling {:interval-ms (Integer/parseInt (get-env "POLLING_INTERVAL_MS" "20000"))}
+   :admin {:username (get-env "ADMIN_USERNAME")
+           :password (get-env "ADMIN_PASSWORD")
+           :session-ttl-ms (Long/parseLong (get-env "ADMIN_SESSION_TTL_MS" "28800000"))} ; 8h
    :environment (keyword (get-env "ENVIRONMENT" "development"))})
 
 (def config